selection-scraper/requirements.txt
selenium>=4.15.0
pandas>=2.0.0
webdriver-manager>=4.0.0
qdrant-client>=1.7.0
openai>=1.0.0
numpy>=1.24.0
pydash>=8.0.0
snowflake-connector-python[secure-local-storage]>=4.0.0

